Split off progress logger
